requirements.txt
# Flask核心
Flask==2.3.3
Flask-CORS==4.0.0

# AI模型
openai==1.3.5
anthropic==0.7.7

# 数据处理
pandas==2.1.3
numpy==1.24.3
requests==2.31.0

# 数据库
mysql-connector-python==8.2.0

# 缓存和性能
redis==5.0.1
python-dotenv==1.0.0

# 工具库
Jinja2==3.1.2
python-dateutil==2.8.2
pytz==2023.3

# 开发和测试
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# 日志和监控
structlog==23.2.0